+++ proxy/docker.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"strconv"
 
+	"github.com/WJQSERVER-STUDIO/go-utils/limitreader"
 	"github.com/cloudwego/hertz/pkg/app"
 )
 
@@ -37,11 +38,22 @@ func GhcrRequest(ctx context.Context, c *app.RequestContext, u string, cfg *conf
 		err    error
 	)
 
+	go func() {
+		<-ctx.Done()
+		if resp != nil && resp.Body != nil {
+			resp.Body.Close()
+		}
+		if req != nil {
+			req.Body.Close()
+		}
+	}()
+
 	method = c.Request.Method()
 
 	rb := client.NewRequestBuilder(string(method), u)
 	rb.NoDefaultHeaders()
 	rb.SetBody(c.Request.BodyStream())
+	rb.WithContext(ctx)
 
 	//req, err = client.NewRequest(string(method), u, c.Request.BodyStream())
 	req, err = rb.Build()
@@ -106,10 +118,16 @@ func GhcrRequest(ctx context.Context, c *app.RequestContext, u string, cfg *conf
 
 	c.Status(resp.StatusCode)
 
+	bodyReader := resp.Body
+
+	if cfg.RateLimit.BandwidthLimit.Enabled {
+		bodyReader = limitreader.NewRateLimitedReader(bodyReader, bandwidthLimit, int(bandwidthBurst), ctx)
+	}
+
 	if contentLength != "" {
-		c.SetBodyStream(resp.Body, bodySize)
+		c.SetBodyStream(bodyReader, bodySize)
 		return
 	}
-	c.SetBodyStream(resp.Body, -1)
+	c.SetBodyStream(bodyReader, -1)
 
 }